Revised bamboo project model and updated dependencies
Downgraded various dependencies to ensure compatibility with the provider-bitbucket module. Updated 'AgentAssignmentResource' and introduced AgentAssignmentModel to reflect changes in the Bamboo client API. Made 'description' field in project schema optional and computed with a default value to accommodate the changes in the API. Adjusted handling of repositories in the deployment model based on the deploymentRepositoryIDs value.