Comment out "id" attribute in Bamboo agent assignment
The "id" attribute in both resource_bamboo_agent_assignment.go and agent_assignment_model.go files has been commented out. This indicates that its usage in defining the assignment of an executable to a Bamboo agent is currently being reconsidered or remodelled. Further modifications to accommodate this change might be needed in the future.