Update terraform-provider-commons to v1.0.3
The 'terraform-provider-commons' has been updated from version v1.0.2 to version v1.0.3 in the 'go.mod' and 'go.sum' files. This update aims to enhance the handling of attribute changes with better precision. It now includes a 'createSlug' modifier in 'resource_bitbucket_repository.go' that triggers resource recreation in Terraform when the repository name is altered.